Ændret fra short til int

+++ EKG_Project/Files/RaspberryPi/IECGSensor.cs
@@ -2,5 +2,5 @@ namespace EKG_Project;
 
 public interface IECGSensor
 {
-    short ReadRawSample();
+    int ReadRawSample();
 }
